Avoid overflow errors
I was getting overflow errors in computing the acceptance probability. If this is so large, it's presumably not a problem (as it's over 1) so we ignore it. It's possible it's actually a symptom of a problem elsewhere, but this won't affect that.

Also made the sampling very slightly more efficient by cacheing likelihoods.