Fixed bug in acceptance prob
Fixed a problem with the computation of the acceptance probability. This is wrong in the original Java implementation: log probabilities are used instead of actual probabilities.

Also fixed a bug of my own in the sampling.

Added a bit more diagnostic output together with topics after each iteration.

Added an option to limit the number of words considered when formatting topics. With a large vocabulary, the formatting can take a long time without this.