Removed constraint on zero-count topics
Had added a constraint to avoid unsampling the last allocation of a word to any topic, to avoid zero counts. However, there's nothing wrong with zero counts. (I thought this was causing numerical problems, but now I think they were due to other bugs, now fixed.)

Removed this constraint, so now a topic can end up completely unused by the sampler if it sees fit. This is the intended behaviour and follows the original implementation.