Allow words to be given as vectors
Allow words in a document for inference to be specified as explicit
vectors. This allows words not in the training vocabulary to be analysed
by supplying new vectors. E.g. if you use fastText vectors, you can
derive a vector from the sub-word info. Or maybe you have vectors from a
bigger set for words that weren't in the training corpus.

Also added oovs_as_nones option to make it easier to match up topics in
the output with the input words.